requirements.txt
fastapi==0.110.0
uvicorn==0.21.1
requests==2.28.1
PyMuPDF==1.23.3
Pillow==10.3.0
chromadb==0.5.5
langchain==0.1.10
transformers==4.43.2
torch==2.0.1

# flash_attn == 2.5.8
# torch == 2.3.1
# accelerate == 0.31.0
# transformers == 4.41.2
